steam_auth/requirements.txt
fastapi
uvicorn
python3-openid
aiohttp